requirements.txt
nltk
langchain_google_genai
langchain_community
pydantic
python-dotenv
streamlit
pypdf